+++ local/ipc/ping.go
@@ -56,6 +56,6 @@ func ping(r pipeline.Router, pingOptions json.RawMessage, style int) {
 	if style == PING_FLOOD {
 		r.OriginateFlooding(po.TTL, dh, d)
 	} else if style == PING_DSR {
-		r.OriginateDSR(po.Destination, uint32(0), dh, d)
+		r.OriginateDSR(po.Destination, int64(0), dh, d)
 	}
 }
